requirements.txt
Pillow 
django
djangorestframework
psycopg2-binary
python-decouple